requirements.txt
flask==3.0.3
langchain==0.2.14
langchain-community==0.2.12
langchain-text-splitters==0.2.2
faiss-cpu==1.8.0.post1
pypdf==4.3.1
tiktoken==0.7.0
requests==2.32.3
huggingface_hub==0.34.4
python-dotenv==1.0.1